+++ example/lib/place_building.dart
@@ -39,7 +39,6 @@ class PlaceBuildingBodyState extends State<PlaceBuildingBody> {
 
   void _onMapCreated(MFMapViewController controller) {
     this.controller = controller;
-    this.controller.enable3DMode(true);
   }
 
   void _onBuildingTapped(MFBuildingId id) {
@@ -142,6 +141,7 @@ class PlaceBuildingBodyState extends State<PlaceBuildingBody> {
             height: 500.0,
             child: MFMapView(
               initialCameraPosition: MFCameraPosition(target: _kInitPosition),
+              buildingsEnabled: true,
               buildings: Set<MFBuilding>.of(buildings.values),
               onMapCreated: _onMapCreated,
             ),